Move styled files to ui
