+++ src/components/ui/EntriesList.styled.js
