+++ src/components/ui/Entry.styled.js
